backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.2
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-settings==2.1.0
python-dotenv==1.1.1
PyYAML==6.0.2
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# Google APIs
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
# Background Tasks
celery==5.3.4
redis==5.0.1
# HTTP Client
httpx==0.25.2
# Encryption
cryptography>=41.0.0
# Date/Time
python-dateutil==2.8.2
